requirements.txt
autopep8==2.3.2
pytest==8.3.5
ipykernel==6.29.5
pandas==2.2.3
scikit-learn==1.6.1
pytest==8.3.5
dash==3.0.4
dash-bootstrap-components==2.0.2
matplotlib==3.10.1
torch==2.7.0+cu128
torchvision-0.22.0